Find Square Root w/out Math.sqrt()
This is my first project(!!) and I am finding the square root of an inputted number without using the sqrt() method from Java's Math library. Something to note is that this is not just a method but a whole little program of it's own, and as such, I have not coded the method that does the square rooting to accept arguments or return values, but the idea is still the same and it can easily be tweaked to do so. Enjoy!